move touch_soon
